Update content for threadly
Here we add the posts and content starting point for the threadly site on the new theme / design.
This includes javadocs not just for threadly but now other artifacts too.
Here we remove a lot of things not needed for our site.
Bump rubyzip from 1.2.3 to 2.2.0 for security.
Bump nokogiri from 1.10.3 to 1.10 for security.8